+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Issue #68: KPI for Gaussian policy
 - Issue #69: Lander with crash
 - Issue #76: Cleaning up priority sweeping model by priority
+- Issue #78: int array key for key generator is not unique 
 
 ## [0.1.0] - 2019-01-28